Moved PaClair class to own file for capability using as the module (not only console)
